+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Input gets an `autofocus` attribute. If true, the input will receive focus on page load
 ### Changed
 ### Fixed